make script more robust

+++ aws/lightsail/scale/jenkins-build-server.sh
@@ -25,34 +25,38 @@ cd /usr/local
 export SDKMAN_DIR="/usr/local/sdkman"
 curl -s "https://get.sdkman.io" | bash
 sleep 10
-cd $SDKMAN_DIR
-pwd
-mkdir etc
-chmod 644 etc
-cat "sdkman_auto_answer=true" >> "$SDKMAN_DIR/etc/config"
-chmod 644 etc/config
-source ./bin/sdkman-init.sh
+if [ -d $SDKMAN_DIR ]; then
+    cd $SDKMAN_DIR
+    mkdir etc
+    chmod 644 etc
+    echo "sdkman_auto_answer=true" >> "$SDKMAN_DIR/etc/config"
+    chmod 644 etc/config
 
-sdk version
-sdk install groovy
-sdk install maven
-sdk install gradle
-sdk install scala
-sdk install sbt
-sdk install ant
-sdk install kotlin
-sdk install kscript
+    [[ -s "$SDKMAN_DIR/bin/sdkman-init.sh" ]] && source "$SDKMAN_DIR/bin/sdkman-init.sh"
 
-# Symlinks to tools
-cd /usr/local
-ln -s sdkman/candidates/maven maven
-ln -s sdkman/candidates/ant ant
-ln -s sdkman/candidates/groovy groovy
-ln -s sdkman/candidates/gradle gradle
-ln -s sdkman/candidates/sbt sbt
-ln -s sdkman/candidates/scala scala
-ln -s sdkman/candidates/kotlin kotlin
-ln -s sdkman/candidates/kscript kscript
+    sdk version
+    sdk install groovy
+    sdk install maven
+    sdk install gradle
+    sdk install scala
+    sdk install sbt
+    sdk install ant
+    sdk install kotlin
+    sdk install kscript
+
+    # Symlinks to tools
+    cd /usr/local
+    ln -s sdkman/candidates/maven maven
+    ln -s sdkman/candidates/ant ant
+    ln -s sdkman/candidates/groovy groovy
+    ln -s sdkman/candidates/gradle gradle
+    ln -s sdkman/candidates/sbt sbt
+    ln -s sdkman/candidates/scala scala
+    ln -s sdkman/candidates/kotlin kotlin
+    ln -s sdkman/candidates/kscript kscript
+else
+  echo "Unable to find $SDKMAN_DIR; several tools not installed"
+fi
 
 # Node via NVM
 cd /usr/local
@@ -61,22 +65,29 @@ chmod 755 nvm
 export NVM_DIR="/usr/local/nvm"
 curl -o- https://raw.githubusercontent.com/creationix/nvm/v0.33.11/install.sh | bash
 sleep 10
-[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
-nvm install node
-nvm install --lts
 
-nvm exec --lts npm install -g karma grunt-cli webpack gulp-cli
-nvm exec npm install -g karma grunt-cli webpack gulp-cli
+if [ -s "$NVM_DIR/nvm.sh" ]; then
+    cd $NVM_DIR
+    ./nvm.sh
+    nvm install node
+    nvm install --lts
+
+    nvm exec --lts npm install -g karma grunt-cli webpack gulp-cli
+    nvm exec npm install -g karma grunt-cli webpack gulp-cli
+fi
 
 # Setup Jenkins user
 cd
 adduser --disabled-password --gecos "" jenkins
 adduser jenkins sudo
-cd /home/jenkins
-mkdir .ssh
-chmod 700 .ssh
-chown jenkins.jenkins .ssh
-cd .ssh
-cp /home/ubuntu/.ssh/authorized_keys .
-chmod 600 authorized_keys
-chown jenkins.jenkins authorized_keys
+
+if [ -d /home/jenkins ]; then
+    cd /home/jenkins
+    mkdir .ssh
+    chmod 700 .ssh
+    chown jenkins.jenkins .ssh
+    cd .ssh
+    cp /home/ubuntu/.ssh/authorized_keys .
+    chmod 600 authorized_keys
+    chown jenkins.jenkins authorized_keys
+fi
